feat: noticbox component
